Crate parity_rpc

OpenEthereum JSON-RPC Servers (WS, HTTP, IPC).

Re-exports

pub extern crate jsonrpc_ws_server as ws;

pub use http::hyper;

pub use v1::extractors::RpcExtractor;

pub use v1::extractors::WsDispatcher;

pub use v1::extractors::WsExtractor;

pub use v1::extractors::WsStats;

pub use v1::informant;

pub use v1::signer;

pub use v1::Metadata;

Modules

dispatch

Utilities and helpers for transaction dispatch.

tests

RPC integration tests.

v1

Ethcore rpc v1.

Structs

AuthCodes

Manages authorization codes for SignerUIs

Host

Host type

IpcRequestContext

Request context

IpcServer

IPC Server handle

NetworkSettings

Networking & RPC settings

PubSubSession

RPC client session Keeps track of active subscriptions and unsubscribes from them upon dropping.

Enums

AccessControlAllowHeaders

Headers allowed to access

AccessControlAllowOrigin

Origins allowed to access

DomainsValidation

Specifies if domains should be validated.

Origin

RPC request origin

RequestMiddlewareAction

Action undertaken by a middleware.

Traits

IpcMetaExtractor

Metadata extractor (per session)

RequestMiddleware

Allows to intercept request and handle it differently.

TimeProvider

Providing current time in seconds

Functions

is_major_importing

Check if client is during major sync or during block import.

is_major_importing_or_waiting

Check if client is during major sync or during block import and allows defining whether ‘waiting for peers’ should be considered a syncing state.

Type Definitions

HttpServer

RPC HTTP Server instance