Module parity_rpc::v1

Ethcore rpc v1.

Compliant with ethereum rpc.

Re-exports

pub use self::extractors::RpcExtractor;

pub use self::extractors::WsDispatcher;

pub use self::extractors::WsExtractor;

pub use self::extractors::WsStats;

pub use self::metadata::Metadata;

Modules

block_import

Block import analysis functions.

dispatch

Utilities and helpers for transaction dispatch.

extractors

OpenEthereum-specific metadata extractors.

informant

RPC Requests Statistics

metadata

OpenEthereum RPC requests Metadata.

signer

Signer utilities

traits

Ethereum rpc interfaces.

Structs

DebugClient

Debug rpc implementation.

EthClient

Eth rpc implementation.

EthClientOptions

Eth RPC options

EthFilterClient

Eth filter rpc implementation for a full node.

EthPubSubClient

Eth PubSub implementation.

NetClient

Net rpc implementation.

NetworkSettings

Networking & RPC settings

ParityAccountsClient

Account management (personal) rpc implementation.

ParityClient

Parity implementation.

ParitySetAccountsClient

Parity-specific account-touching RPC interfaces.

ParitySetClient

Parity-specific rpc interface for operations altering the settings.

PersonalClient

Account management (personal) rpc implementation.

PubSubClient

Parity PubSub implementation.

RpcClient

RPC generic methods implementation.

SecretStoreClient

Parity implementation.

SignerClient

Transactions confirmation (personal) rpc implementation.

SigningQueueClient

Implementation of functions that require signing when no trusted signer is used.

SigningUnsafeClient

Implementation of functions that require signing when no trusted signer is used.

TracesClient

Traces api implementation.

Web3Client

Web3 rpc implementation.

Enums

Origin

RPC request origin

Traits

Debug

Debug RPC interface.

Eth

Eth rpc interface.

EthFilter

Eth filters rpc api (polling).

EthPubSub

Eth PUB-SUB rpc interface.

EthSigning

Signing methods implementation relying on unlocked accounts.

Net

Net rpc interface.

Parity

Parity-specific rpc interface.

ParityAccounts

Personal Parity rpc interface.

ParityAccountsInfo

Parity-specific read-only accounts rpc interface.

ParitySet

Parity-specific rpc interface for operations altering the settings.

ParitySetAccounts

Parity-specific rpc interface for operations altering the account-related settings.

ParitySigning

Signing methods implementation.

Personal

Personal rpc interface. Safe (read-only) functions.

PubSub

Parity-specific PUB-SUB rpc interface.

Rpc

RPC Interface.

SecretStore

Parity-specific rpc interface.

Signer

Signer extension for confirmations rpc interface.

Traces

Traces specific rpc interface.

Web3

Web3 rpc interface.