Module parity_rpc::v1::dispatch

Utilities and helpers for transaction dispatch.

Structs

FullDispatcher

A dispatcher which uses references to a client and miner in order to sign requests locally.

Reservations

Manages currently reserved and prospective nonces for multiple senders.

Signer

Account-aware signer

Enums

SignMessage

Payload to sign

SignWith

Values used to unlock accounts for signing.

WithToken

A value, potentially accompanied by a signing token.

Traits

Accounts

Abstract transaction signer.

Dispatcher

Has the capability to dispatch, sign, and decrypt.

PostSign

action to execute after signing e.g importing a transaction into the chain

Functions

default_gas_price

Extract the default gas price from a client and miner.

default_max_priority_fee_per_gas

Extract the default priority gas price from a client and miner.

eth_data_hash

Returns a eth_sign-compatible hash of data to sign. The data is prepended with special message to prevent malicious DApps from using the function to sign forged transactions.

execute

Execute a confirmation payload.

from_rpc

Convert RPC confirmation payload to signer confirmation payload. May need to resolve in the future to fetch things like gas price.

Type Definitions

AccountToken

Single-use account token.