Trait parity_rpc::v1::EthSigning

pub trait EthSigning: Sized + Send + Sync + 'static {
    type Metadata: Metadata;
    fn sign(&self, _: Self::Metadata, _: H160, _: Bytes) -> BoxFuture<H520>;

    fn send_transaction(
        &self, 
        _: Self::Metadata, 
        _: TransactionRequest
    ) -> BoxFuture<H256>;

    fn sign_transaction(
        &self, 
        _: Self::Metadata, 
        _: TransactionRequest
    ) -> BoxFuture<RichRawTransaction>;

    fn to_delegate(self) -> IoDelegate<Self, Self::Metadata> { ... }
}

Signing methods implementation relying on unlocked accounts.

Associated Types

type Metadata: Metadata

RPC Metadata

Required methods

fn sign(&self, _: Self::Metadata, _: H160, _: Bytes) -> BoxFuture<H520>

Signs the hash of data with given address signature.

fn send_transaction(
&self,
_: Self::Metadata,
_: TransactionRequest
) -> BoxFuture<H256>

Sends transaction; will block waiting for signer to return the transaction hash. If Signer is disable it will require the account to be unlocked.

fn sign_transaction(
&self,
_: Self::Metadata,
_: TransactionRequest
) -> BoxFuture<RichRawTransaction>

Signs transactions without dispatching it to the network. Returns signed transaction RLP representation and the transaction itself. It can be later submitted using eth_sendRawTransaction/eth_submitTransaction.

Provided methods

fn to_delegate(self) -> IoDelegate<Self, Self::Metadata>

Create an IoDelegate, wiring rpc calls to the trait methods.