Trait parity_rpc::v1::Personal

pub trait Personal: Sized + Send + Sync + 'static {
    type Metadata: Metadata;
    fn accounts(&self) -> Result<Vec<H160>>;

    fn new_account(&self, _: String) -> Result<H160>;

    fn unlock_account(
        &self, 
        _: H160, 
        _: String, 
        _: Option<U128>
    ) -> Result<bool>;

    fn sign(&self, _: Bytes, _: H160, _: String) -> BoxFuture<H520>;

    fn sign_typed_data(&self, _: EIP712, _: H160, _: String) -> BoxFuture<H520>;

    fn sign_191(
        &self, 
        _: EIP191Version, 
        _: Value, 
        _: H160, 
        _: String
    ) -> BoxFuture<H520>;

    fn ec_recover(&self, _: Bytes, _: H520) -> BoxFuture<H160>;

    fn sign_transaction(
        &self, 
        _: Self::Metadata, 
        _: TransactionRequest, 
        _: String
    ) -> BoxFuture<RpcRichRawTransaction>;

    fn send_transaction(
        &self, 
        _: Self::Metadata, 
        _: TransactionRequest, 
        _: String
    ) -> BoxFuture<H256>;

    fn sign_and_send_transaction(
        &self, 
        _: Self::Metadata, 
        _: TransactionRequest, 
        _: String
    ) -> BoxFuture<H256>;

    fn to_delegate(self) -> IoDelegate<Self, Self::Metadata> { ... }
}

Personal rpc interface. Safe (read-only) functions.

Associated Types

type Metadata: Metadata

RPC Metadata

Required methods

fn accounts(&self) -> Result<Vec<H160>>

Lists all stored accounts

fn new_account(&self, _: String) -> Result<H160>

Creates new account (it becomes new current unlocked account) Param is the password for the account.

fn unlock_account(&self, _: H160, _: String, _: Option<U128>) -> Result<bool>

Unlocks specified account for use (can only be one unlocked account at one moment)

fn sign(&self, _: Bytes, _: H160, _: String) -> BoxFuture<H520>

Signs the hash of data with given account signature using the given password to unlock the account during the request.

fn sign_typed_data(&self, _: EIP712, _: H160, _: String) -> BoxFuture<H520>

Produces an EIP-712 compliant signature with given account using the given password to unlock the account during the request.

fn sign_191(
&self,
_: EIP191Version,
_: Value,
_: H160,
_: String
) -> BoxFuture<H520>

Signs an arbitrary message based on the version specified

fn ec_recover(&self, _: Bytes, _: H520) -> BoxFuture<H160>

Returns the account associated with the private key that was used to calculate the signature in personal_sign.

fn sign_transaction(
&self,
_: Self::Metadata,
_: TransactionRequest,
_: String
) -> BoxFuture<RpcRichRawTransaction>

Signs transaction. The account is not unlocked in such case.

fn send_transaction(
&self,
_: Self::Metadata,
_: TransactionRequest,
_: String
) -> BoxFuture<H256>

Sends transaction and signs it in single call. The account is not unlocked in such case.

fn sign_and_send_transaction(
&self,
_: Self::Metadata,
_: TransactionRequest,
_: String
) -> BoxFuture<H256>

@deprecated alias for personal_sendTransaction.

Provided methods

fn to_delegate(self) -> IoDelegate<Self, Self::Metadata>

Create an IoDelegate, wiring rpc calls to the trait methods.