Trait parity_rpc::v1::Traces

pub trait Traces: Sized + Send + Sync + 'static {
    fn filter(&self, _: TraceFilter) -> Result<Option<Vec<LocalizedTrace>>>;

    fn trace(&self, _: H256, _: Vec<Index>) -> Result<Option<LocalizedTrace>>;

    fn transaction_traces(&self, _: H256) -> Result<Option<Vec<LocalizedTrace>>>;

    fn block_traces(
        &self, 
        _: BlockNumber
    ) -> Result<Option<Vec<LocalizedTrace>>>;

    fn call(
        &self, 
        _: CallRequest, 
        _: Vec<String>, 
        _: Option<BlockNumber>
    ) -> Result<TraceResults>;

    fn call_many(
        &self, 
        _: Vec<(CallRequest, Vec<String>)>, 
        _: Option<BlockNumber>
    ) -> Result<Vec<TraceResults>>;

    fn raw_transaction(
        &self, 
        _: Bytes, 
        _: Vec<String>, 
        _: Option<BlockNumber>
    ) -> Result<TraceResults>;

    fn replay_transaction(
        &self, 
        _: H256, 
        _: Vec<String>
    ) -> Result<TraceResults>;

    fn replay_block_transactions(
        &self, 
        _: BlockNumber, 
        _: Vec<String>
    ) -> Result<Vec<TraceResultsWithTransactionHash>>;

    fn to_delegate<M: Metadata>(self) -> IoDelegate<Self, M> { ... }
}

Traces specific rpc interface.

Required methods

fn filter(&self, _: TraceFilter) -> Result<Option<Vec<LocalizedTrace>>>

Returns traces matching given filter.

fn trace(&self, _: H256, _: Vec<Index>) -> Result<Option<LocalizedTrace>>

Returns transaction trace at given index.

fn transaction_traces(&self, _: H256) -> Result<Option<Vec<LocalizedTrace>>>

Returns all traces of given transaction.

fn block_traces(&self, _: BlockNumber) -> Result<Option<Vec<LocalizedTrace>>>

Returns all traces produced at given block.

fn call(
&self,
_: CallRequest,
_: Vec<String>,
_: Option<BlockNumber>
) -> Result<TraceResults>

Executes the given call and returns a number of possible traces for it.

fn call_many(
&self,
_: Vec<(CallRequest, Vec<String>)>,
_: Option<BlockNumber>
) -> Result<Vec<TraceResults>>

Executes all given calls and returns a number of possible traces for each of it.

fn raw_transaction(
&self,
_: Bytes,
_: Vec<String>,
_: Option<BlockNumber>
) -> Result<TraceResults>

Executes the given raw transaction and returns a number of possible traces for it.

fn replay_transaction(&self, _: H256, _: Vec<String>) -> Result<TraceResults>

Executes the transaction with the given hash and returns a number of possible traces for it.

fn replay_block_transactions(
&self,
_: BlockNumber,
_: Vec<String>
) -> Result<Vec<TraceResultsWithTransactionHash>>

Executes all the transactions at the given block and returns a number of possible traces for each transaction.

Provided methods

fn to_delegate<M: Metadata>(self) -> IoDelegate<Self, M>

Create an IoDelegate, wiring rpc calls to the trait methods.