Trait parity_rpc::v1::traits::parity_set::ParitySet

pub trait ParitySet: Sized + Send + Sync + 'static {

    fn set_min_gas_price(&self, _: U256) -> Result<bool>;

    fn set_gas_floor_target(&self, _: U256) -> Result<bool>;

    fn set_gas_ceil_target(&self, _: U256) -> Result<bool>;

    fn set_extra_data(&self, _: Bytes) -> Result<bool>;

    fn set_author(&self, _: H160) -> Result<bool>;

    fn set_engine_signer_secret(&self, _: H256) -> Result<bool>;

    fn clear_engine_signer(&self) -> Result<bool>;

    fn set_transactions_limit(&self, _: usize) -> Result<bool>;

    fn set_tx_gas_limit(&self, _: U256) -> Result<bool>;

    fn add_reserved_peer(&self, _: String) -> Result<bool>;

    fn remove_reserved_peer(&self, _: String) -> Result<bool>;

    fn drop_non_reserved_peers(&self) -> Result<bool>;

    fn accept_non_reserved_peers(&self) -> Result<bool>;

    fn start_network(&self) -> Result<bool>;

    fn stop_network(&self) -> Result<bool>;

    fn set_mode(&self, _: String) -> Result<bool>;

    fn set_spec_name(&self, _: String) -> Result<bool>;

    fn hash_content(&self, _: String) -> BoxFuture<H256>;

    fn remove_transaction(&self, _: H256) -> Result<Option<Transaction>>;

    fn to_delegate<M: Metadata>(self) -> IoDelegate<Self, M> { ... }

}

Parity-specific rpc interface for operations altering the settings.

Required methods

fn set_min_gas_price(&self, _: U256) -> Result<bool>

Sets new minimal gas price for mined blocks.

fn set_gas_floor_target(&self, _: U256) -> Result<bool>

Sets new gas floor target for mined blocks.

fn set_gas_ceil_target(&self, _: U256) -> Result<bool>

Sets new gas ceiling target for mined blocks.

fn set_extra_data(&self, _: Bytes) -> Result<bool>

Sets new extra data for mined blocks.

fn set_author(&self, _: H160) -> Result<bool>

Sets new author for mined block.

fn set_engine_signer_secret(&self, _: H256) -> Result<bool>

Sets the secret of engine signer account.

fn clear_engine_signer(&self) -> Result<bool>

Unsets the engine signer account address.

fn set_transactions_limit(&self, _: usize) -> Result<bool>

Sets the limits for transaction queue.

fn set_tx_gas_limit(&self, _: U256) -> Result<bool>

Sets the maximum amount of gas a single transaction may consume.

fn add_reserved_peer(&self, _: String) -> Result<bool>

Add a reserved peer.

fn remove_reserved_peer(&self, _: String) -> Result<bool>

Remove a reserved peer.

fn drop_non_reserved_peers(&self) -> Result<bool>

Drop all non-reserved peers.

fn accept_non_reserved_peers(&self) -> Result<bool>

Accept non-reserved peers (default behavior)

fn start_network(&self) -> Result<bool>

Start the network.

@deprecated - Use set_mode("active") instead.

fn stop_network(&self) -> Result<bool>

Stop the network.

@deprecated - Use set_mode("offline") instead.

fn set_mode(&self, _: String) -> Result<bool>

Set the mode. Argument must be one of: “active”, “passive”, “dark”, “offline”.

fn set_spec_name(&self, _: String) -> Result<bool>

Set the network spec. Argument must be one of pre-configured chains or a filename.

fn hash_content(&self, _: String) -> BoxFuture<H256>

Hash a file content under given URL.

fn remove_transaction(&self, _: H256) -> Result<Option<Transaction>>

Removes transaction from transaction queue. Makes sense only for transactions that were not propagated to other peers yet like scheduled transactions or transactions in future. It might also work for some local transactions with to low gas price or excessive gas limit that are not accepted by other peers whp. Returns true when transaction was removed, false if it was not found.

Provided methods

fn to_delegate<M: Metadata>(self) -> IoDelegate<Self, M>

Create an IoDelegate, wiring rpc calls to the trait methods.