Trait parity_rpc::v1::traits::ParityAccounts

pub trait ParityAccounts: Sized + Send + Sync + 'static {

    fn all_accounts_info(&self) -> Result<BTreeMap<H160, ExtAccountInfo>>;

    fn new_account_from_phrase(&self, _: String, _: Password) -> Result<H160>;

    fn new_account_from_wallet(&self, _: String, _: Password) -> Result<H160>;

    fn new_account_from_secret(&self, _: H256, _: Password) -> Result<H160>;

    fn test_password(&self, _: H160, _: Password) -> Result<bool>;

    fn change_password(&self, _: H160, _: Password, _: Password) -> Result<bool>;

    fn kill_account(&self, _: H160, _: Password) -> Result<bool>;

    fn remove_address(&self, _: H160) -> Result<bool>;

    fn set_account_name(&self, _: H160, _: String) -> Result<bool>;

    fn set_account_meta(&self, _: H160, _: String) -> Result<bool>;

    fn create_vault(&self, _: String, _: Password) -> Result<bool>;

    fn open_vault(&self, _: String, _: Password) -> Result<bool>;

    fn close_vault(&self, _: String) -> Result<bool>;

    fn list_vaults(&self) -> Result<Vec<String>>;

    fn list_opened_vaults(&self) -> Result<Vec<String>>;

    fn change_vault_password(&self, _: String, _: Password) -> Result<bool>;

    fn change_vault(&self, _: H160, _: String) -> Result<bool>;

    fn get_vault_meta(&self, _: String) -> Result<String>;

    fn set_vault_meta(&self, _: String, _: String) -> Result<bool>;

    fn derive_key_hash(
        &self, 
        _: H160, 
        _: Password, 
        _: DeriveHash, 
        _: bool
    ) -> Result<H160>;

    fn derive_key_index(
        &self, 
        _: H160, 
        _: Password, 
        _: Vec<DeriveHierarchicalItem>, 
        _: bool
    ) -> Result<H160>;

    fn export_account(&self, _: H160, _: Password) -> Result<KeyFile>;

    fn sign_message(&self, _: H160, _: Password, _: H256) -> Result<H520>;

    fn to_delegate<M: Metadata>(self) -> IoDelegate<Self, M> { ... }

}

Personal Parity rpc interface.

Required methods

fn all_accounts_info(&self) -> Result<BTreeMap<H160, ExtAccountInfo>>

Returns accounts information.

fn new_account_from_phrase(&self, _: String, _: Password) -> Result<H160>

Creates new account from the given phrase using standard brainwallet mechanism. Second parameter is password for the new account.

fn new_account_from_wallet(&self, _: String, _: Password) -> Result<H160>

Creates new account from the given JSON wallet. Second parameter is password for the wallet and the new account.

fn new_account_from_secret(&self, _: H256, _: Password) -> Result<H160>

Creates new account from the given raw secret. Second parameter is password for the new account.

fn test_password(&self, _: H160, _: Password) -> Result<bool>

Returns true if given password would unlock given account. Arguments: account, password.

fn change_password(&self, _: H160, _: Password, _: Password) -> Result<bool>

Changes an account’s password. Arguments: account, password, new_password.

fn kill_account(&self, _: H160, _: Password) -> Result<bool>

Permanently deletes an account. Arguments: account, password.

fn remove_address(&self, _: H160) -> Result<bool>

Permanently deletes an address from the addressbook Arguments: address

fn set_account_name(&self, _: H160, _: String) -> Result<bool>

Set an account’s name.

fn set_account_meta(&self, _: H160, _: String) -> Result<bool>

Set an account’s metadata string.

fn create_vault(&self, _: String, _: Password) -> Result<bool>

Create new vault.

fn open_vault(&self, _: String, _: Password) -> Result<bool>

Open existing vault.

fn close_vault(&self, _: String) -> Result<bool>

Close previously opened vault.

fn list_vaults(&self) -> Result<Vec<String>>

List all vaults.

fn list_opened_vaults(&self) -> Result<Vec<String>>

List all currently opened vaults.

fn change_vault_password(&self, _: String, _: Password) -> Result<bool>

Change vault password.

fn change_vault(&self, _: H160, _: String) -> Result<bool>

Change vault of the given address.

fn get_vault_meta(&self, _: String) -> Result<String>

Get vault metadata string.

fn set_vault_meta(&self, _: String, _: String) -> Result<bool>

Set vault metadata string.

fn derive_key_hash(
&self,
_: H160,
_: Password,
_: DeriveHash,
_: bool
) -> Result<H160>

Derive new address from given account address using specific hash. Resulting address can be either saved as a new account (with the same password).

fn derive_key_index(
&self,
_: H160,
_: Password,
_: Vec<DeriveHierarchicalItem>,
_: bool
) -> Result<H160>

Derive new address from given account address using hierarchical derivation (sequence of 32-bit integer indices). Resulting address can be either saved as a new account (with the same password).

fn export_account(&self, _: H160, _: Password) -> Result<KeyFile>

Exports an account with given address if provided password matches.

fn sign_message(&self, _: H160, _: Password, _: H256) -> Result<H520>

Sign raw hash with the key corresponding to address and password.

Provided methods

fn to_delegate<M: Metadata>(self) -> IoDelegate<Self, M>

Create an IoDelegate, wiring rpc calls to the trait methods.